contract-obligations-agent/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
gradio>=4.44
pydantic>=2.8
chromadb>=0.5
qdrant-client>=1.11
litellm>=1.44
SQLAlchemy>=2.0
python-docx>=1.1
pypdf>=5.0
pytest>=8.0
httpx>=0.27
python-multipart>=0.0.9

